replace data state variable on dataManager with parseData disk read
