+++ src/dataManager.ts
@@ -5,69 +5,98 @@ import { DEV_BUILD } from './config';
 import { getCurrentDate, createMockData } from './utils'
 import { TFile } from 'obsidian';
 import { METRIC_TYPES } from './constants';
+import { isActivityHeatmapData } from './utils';
 
 /**
  * Manages activity heatmap data for the plugin.
  */
 export class ActivityHeatmapDataManager {
     private metricManager: MetricManager;
-    private data: ActivityHeatmapData;
-
 
     /**
-     * Creates an instance of ActivityHeatmapDataManager.
+     * Creates an instance of ActivityHeatmapDataManager, which is responsible for reads/writes of activity heatmap data to disk.
      * @param plugin - The main plugin instance.
-     * @param loadedData - The parsed activity heatmap data from the data.json file.
      */
-    constructor(private plugin: ActivityHeatmapPlugin, loadedData: ActivityHeatmapData) {
+    constructor(private plugin: ActivityHeatmapPlugin) {
         this.metricManager = new MetricManager(plugin);
-        this.data = loadedData;
     }
 
     /**
      * Updates metrics for a single file and saves the data.
-     * This method is more efficient than updating all files as it only processes
-     * the changes for a single file.
-     * 
      * @param file - The Obsidian TFile to update metrics for
      */
     async updateMetricsForFile(file: TFile) {
         const today = getCurrentDate();
+        const data = await this.parseActivityData();
 
         for (const metricType of METRIC_TYPES) {
             const { checkpoint, activity } = await this.metricManager.calculateMetricsForFile(
                 metricType,
                 file,
-                this.data,
+                data,
                 today
             );
             
-            // Update only the specific file's checkpoint
-            this.data.checkpoints[metricType] = {
-                ...this.data.checkpoints[metricType],
+            data.checkpoints[metricType] = {
+                ...data.checkpoints[metricType],
                 [file.path]: checkpoint[file.path]
             };
-            this.data.activityOverTime[metricType] = activity;
+            data.activityOverTime[metricType] = activity;
         }
 
-        await this.plugin.saveData(this.data);
+        await this.plugin.saveData(data);
     }
 
     /**
      * Removes metrics for a deleted file from all metric types.
-     * This cleanup ensures we don't maintain data for files that no longer exist.
-     * 
      * @param filePath - The path of the file that was deleted
      */
     async removeFileMetrics(filePath: string) {
+        const data = await this.parseActivityData();
         for (const metricType of METRIC_TYPES) {
-            if (this.data.checkpoints[metricType]) {
-                delete this.data.checkpoints[metricType][filePath];
+            if (data.checkpoints[metricType]) {
+                delete data.checkpoints[metricType][filePath];
             }
         }
 
-        await this.plugin.saveData(this.data);
+        await this.plugin.saveData(data);
     }
+
+    /**
+	 * Parses the data read from disk (plugin's data.json file) into a usable format.
+     * If the data is not in the expected format, it returns an empty data structure to prevent errors.
+	 * @returns The parsed activity data.
+	 */
+	async parseActivityData(): Promise<ActivityHeatmapData> {
+		const loadedData = await this.plugin.loadData();
+
+		const emptyFrame: ActivityHeatmapData = {
+			checkpoints: METRIC_TYPES.reduce((acc, metric) => ({
+				...acc,
+				[metric]: {} as Record<string, number>
+			}), {} as Record<MetricType, Record<string, number>>),
+			activityOverTime: METRIC_TYPES.reduce((acc, metric) => ({
+				...acc,
+				[metric]: {} as Record<string, number>
+			}), {} as Record<MetricType, Record<string, number>>)
+		};
+
+		// Case of new user (no data.json)
+		if (!loadedData) {
+			return emptyFrame;
+		}
+
+		// Case of invalid or malformed activity heatmap data
+		if (!isActivityHeatmapData(loadedData)) {
+			return emptyFrame;
+		}
+
+		// Correct case: extract only the ActivityHeatmapData properties
+		return {
+			checkpoints: loadedData.checkpoints,
+			activityOverTime: loadedData.activityOverTime
+		};
+	}
 
     /**
      * Retrieves activity heatmap data for a specific metric type.
@@ -79,7 +108,8 @@ export class ActivityHeatmapDataManager {
             console.log("Using mock data");
             return createMockData();
         }
-        return this.data.activityOverTime[metricType];
+        const data = await this.parseActivityData();
+        return data.activityOverTime[metricType];
     }
 
 }
