+++ src/main.ts
@@ -1,12 +1,11 @@
 import { Plugin } from 'obsidian';
-import type { ActivityHeatmapData, ActivityHeatmapSettings, MetricType } from './types'
+import type { ActivityHeatmapSettings } from './types'
 import { ActivityHeatmapDataManager } from './dataManager'
-import { DEFAULT_SETTINGS, METRIC_TYPES } from './constants'
+import { DEFAULT_SETTINGS } from './constants'
 import { ActivityHeatmapSettingTab } from './settings'
 import { HeatmapModal } from './components/heatmapModal';
 import { DEV_BUILD } from './config';
 import { TFile } from 'obsidian';
-import { isActivityHeatmapData } from './utils';
 
 export default class ActivityHeatmapPlugin extends Plugin {
 	settings: ActivityHeatmapSettings;
@@ -18,7 +17,7 @@ export default class ActivityHeatmapPlugin extends Plugin {
 		await this.loadSettings();
 		this.addSettingTab(new ActivityHeatmapSettingTab(this.app, this));
 
-		this.dataManager = new ActivityHeatmapDataManager(this, await this.parseActivityData());
+		this.dataManager = new ActivityHeatmapDataManager(this);
 
 		this.registerEvent(
 			this.app.vault.on('modify', (file) => {
@@ -81,42 +80,6 @@ export default class ActivityHeatmapPlugin extends Plugin {
 		};
 		await this.saveData(dataToSave);
 	}
-
-	/**
-	 * Parses the activity data from the plugin's data.json file.
-	 * @returns The parsed activity data.
-	 */
-	async parseActivityData(): Promise<ActivityHeatmapData> {
-		const loadedData = await this.loadData();
-
-		const emptyFrame: ActivityHeatmapData = {
-			checkpoints: METRIC_TYPES.reduce((acc, metric) => ({
-				...acc,
-				[metric]: {} as Record<string, number>
-			}), {} as Record<MetricType, Record<string, number>>),
-			activityOverTime: METRIC_TYPES.reduce((acc, metric) => ({
-				...acc,
-				[metric]: {} as Record<string, number>
-			}), {} as Record<MetricType, Record<string, number>>)
-		};
-
-		// Case of new user (no data.json)
-		if (!loadedData) {
-			return emptyFrame;
-		}
-
-		// Case of invalid or malformed activity heatmap data
-		if (!isActivityHeatmapData(loadedData)) {
-			return emptyFrame;
-		}
-
-		// Correct case: extract only the ActivityHeatmapData properties
-		return {
-			checkpoints: loadedData.checkpoints,
-			activityOverTime: loadedData.activityOverTime
-		};
-	}
-
 }
